--- lib/pymedphys/_streamlit/apps/wlutz/__init__.py
@@ -1,1 +1,0 @@
-from .main import main
