+++ lib/pymedphys/_streamlit/index.py
@@ -17,14 +17,15 @@ import time
 
 from pymedphys._imports import streamlit as st
 
-from pymedphys._streamlit.apps import anonymise_monaco as _anonymise_monaco
-from pymedphys._streamlit.apps import dashboard as _dashboard
-from pymedphys._streamlit.apps import electrons as _electrons
-from pymedphys._streamlit.apps import iviewdb as _iviewdb
 from pymedphys._streamlit.apps import metersetmap as _metersetmap
 from pymedphys._streamlit.apps import pseudonymise as _pseudonymise
-from pymedphys._streamlit.apps import wlutz as _wlutz
 from pymedphys._streamlit.utilities import session
+
+from pymedphys._experimental.streamlit.apps import anonymise_monaco as _anonymise_monaco
+from pymedphys._experimental.streamlit.apps import dashboard as _dashboard
+from pymedphys._experimental.streamlit.apps import electrons as _electrons
+from pymedphys._experimental.streamlit.apps import iviewdb as _iviewdb
+from pymedphys._experimental.streamlit.apps import wlutz as _wlutz
 
 HERE = pathlib.Path(__file__).parent.resolve()
 FAVICON = str(HERE.joinpath("pymedphys.png"))
